Abstract FormFieldWrapper to HOC
